# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Companies",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://api.laravel.com"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User Register | register()",
                "description": "Sign in just by entering name and email and password.",
                "operationId": "App\\Http\\Controllers\\AuthController::register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registered User.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "An error has occurred.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User Login | login()",
                "description": "Log in just by entering email and password.",
                "operationId": "App\\Http\\Controllers\\AuthController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logged in user.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "An error has occurred.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User Logout | logout()",
                "description": "Destroy user token by passing the same token.",
                "operationId": "App\\Http\\Controllers\\AuthController::logout",
                "responses": {
                    "200": {
                        "description": "Show all companies."
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/api/companies": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Companies list | index()",
                "description": "See all registered companies.",
                "operationId": "App\\Http\\Controllers\\CompanyController::index",
                "responses": {
                    "200": {
                        "description": "list all companies."
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            },
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Company Create | store()",
                "description": "Register a company simply by defining its name and description.",
                "operationId": "App\\Http\\Controllers\\CompanyController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created company.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "An error has occurred.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/api/companies/{id}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Company Show | show()",
                "description": "See just a registered company.",
                "operationId": "App\\Http\\Controllers\\CompanyController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show a company.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "An error has occurred.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Company"
                ],
                "summary": "Company Update | update()",
                "description": "Update a company simply by defining its name and description and ID.",
                "operationId": "App\\Http\\Controllers\\CompanyController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated company.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "An error has occurred.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Company"
                ],
                "summary": "Company Delete | destroy()",
                "description": "Delete a company simply by defining an ID.",
                "operationId": "App\\Http\\Controllers\\CompanyController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted company.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "An error has occurred.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://api.laravel.com/oauth/authorize",
                        "tokenUrl": "http://api.laravel.com/oauth/token",
                        "refreshUrl": "http://api.laravel.com/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}